delete MainController
break it up into stateless utility functions
move navigation to individual mages
move state to global context
