+++ frontend/src/main.tsx
@@ -1,28 +1,24 @@
-import React, { createContext } from "react";
+import React from "react";
 import { createRoot } from "react-dom/client";
 import PageCoordinator from "./pages/PageCoordinator";
-import MainController from "./MainController";
 import { BrowserRouter } from "react-router-dom";
 import { ThemeProvider } from "@mui/material";
 import { theme } from "./theme";
 import { GlobalStateProvider } from "./GlobalStateContext";
 
 console.log("[ERWT]: Renderer execution started");
-export const AppContext = createContext(null);
 
-const mainController = new MainController();
 // Application to Render
 const app = (
   <ThemeProvider theme={theme}>
     <BrowserRouter>
-      <AppContext.Provider value={mainController as any}>
-        <GlobalStateProvider>
-          <PageCoordinator />
-        </GlobalStateProvider>
-      </AppContext.Provider>
+      <GlobalStateProvider>
+        <PageCoordinator />
+      </GlobalStateProvider>
     </BrowserRouter>
   </ThemeProvider>
 );
 
 // Render application in DOM
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 createRoot(document.getElementById("app") as any).render(app);
