+++ frontend/src/pages/CreateLobby.tsx
@@ -1,20 +1,22 @@
-import React, { useContext, useState } from "react";
+import React, { useState } from "react";
 import "./CreateLobby.scss";
 import { Box, Checkbox, FormControlLabel } from "@mui/material";
-import type MainController from "@src/MainController";
 import Navbar from "@src/components/Navbar";
-import { AppContext } from "@src/main";
 import Wrapper from "@src/components/Wrapper";
 import Button from "@src/components/Button";
 import NumericField from "@src/components/NumericField";
 import { useGlobalStateContext } from "@src/GlobalStateContext";
+import { useNavigate } from "react-router-dom";
+import { Page } from "@src/pages/PageCoordinator";
+import { createLobby } from "@src/services/utils";
 
 const CreateLobby: React.FC = () => {
-  const mainController: MainController = useContext(AppContext) as any;
+  const navigate = useNavigate();
   const {
     selectedNftState: [selectedNft],
     selectedDeckState: [selectedDeck],
     collection,
+    joinedLobbyRawState: [, setJoinedLobbyRaw],
   } = useGlobalStateContext();
 
   const [numberOfRounds, setNumberOfRounds] = useState("5");
@@ -33,7 +35,7 @@ const CreateLobby: React.FC = () => {
     const numberOfRoundsNum = parseInt(numberOfRounds);
     const turnLengthNum = parseInt(turnLength);
 
-    await mainController.createLobby(
+    const createdLobby = await createLobby(
       selectedNft.nft,
       selectedDeck.map((card) => {
         if (collection.cards?.[card] == null)
@@ -48,6 +50,8 @@ const CreateLobby: React.FC = () => {
       isHidden,
       isPractice
     );
+    setJoinedLobbyRaw(createdLobby);
+    navigate(Page.Game);
   };
 
   return (
