+++ frontend/src/utils/index.ts
@@ -1,5 +1,4 @@
 import { format } from "date-fns";
-import React from "react";
 
 export type UseStateResponse<T> = [T, React.Dispatch<React.SetStateAction<T>>];
